+++ src/commands/clone.rs
@@ -1,12 +1,12 @@
 use crate::github;
-use crate::runtime;
+use crate::operations;
 use crate::runtime::Step;
 
 pub fn clone(org: &str) -> Vec<Step> {
     let mut result = vec![];
     let repos = github::get_repos(org);
     for (i, repo) in repos.into_iter().enumerate() {
-        result.push(runtime::clone(i, repo.clone_url));
+        result.push(operations::clone(i, repo.clone_url));
     }
     result
 }
