+++ src/operations/mod.rs
@@ -1,1 +1,3 @@
 mod clone_repo;
+
+pub use clone_repo::clone;
