+++ src/runtime/data.rs
@@ -1,0 +1,15 @@
+use serde::{Deserialize, Serialize};
+use std::fmt::Display;
+
+#[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
+pub struct Step {
+    pub id: usize,
+    pub command: String,
+    pub args: Vec<String>,
+}
+
+impl Display for Step {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{} {} {}", self.id, self.command, self.args.join(" "))
+    }
+}
