+++ src/runtime/mod.rs
@@ -1,5 +1,7 @@
-mod operation;
+mod data;
 mod persistence;
+mod runner;
 
-pub use operation::{clone, run, Step};
+pub use data::Step;
 pub use persistence::{delete, load, save};
+pub use runner::run;
