+++ src/runtime/runner.rs
@@ -1,0 +1,20 @@
+use super::Step;
+use std::process::Command;
+
+/// executes the given steps, returns the not executed steps in case of an issue
+pub fn run(steps: Vec<Step>) -> Vec<Step> {
+    let mut steps_iter = steps.into_iter();
+    while let Some(step) = steps_iter.next() {
+        println!("{}", step);
+        let mut command = Command::new(step.command);
+        command.args(step.args);
+        let success = match command.status() {
+            Ok(status) => status.success(),
+            _ => false,
+        };
+        if !success {
+            return steps_iter.collect();
+        }
+    }
+    vec![]
+}
